requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1

Flask-Login==0.6.3
Werkzeug==3.0.3

openai==1.40.0
